+++ package.json
@@ -33,6 +33,7 @@
     "morgan": "^1.8.1",
     "passport": "^0.3.2",
     "passport-google-oauth": "^1.0.0",
+    "passport-strava-oauth2": "^0.1.3",
     "pg": "^6.1.2",
     "pg-hstore": "^2.3.2",
     "prop-types": "^15.5.8",
@@ -45,7 +46,8 @@
     "redux-logger": "^2.8.1",
     "redux-thunk": "^2.2.0",
     "sequelize": "^4.3.1",
-    "socket.io": "^2.0.3"
+    "socket.io": "^2.0.3",
+    "strava-v3": "^1.14.0"
   },
   "devDependencies": {
     "@babel/core": "^7.0.0-beta.40",
