--- server/auth/google.js
@@ -1,60 +1,0 @@
-const passport = require('passport')
-const router = require('express').Router()
-const GoogleStrategy = require('passport-google-oauth').OAuth2Strategy
-const {User} = require('../db/models')
-module.exports = router
-
-/**
- * For OAuth keys and other secrets, your Node process will search
- * process.env to find environment variables. On your production server,
- * you will be able to set these environment variables with the appropriate
- * values. In development, a good practice is to keep a separate file with
- * these secrets that you only share with your team - it should NOT be tracked
- * by git! In this case, you may use a file called `secrets.js`, which will
- * set these environment variables like so:
- *
- * process.env.GOOGLE_CLIENT_ID = 'your google client id'
- * process.env.GOOGLE_CLIENT_SECRET = 'your google client secret'
- * process.env.GOOGLE_CALLBACK = '/your/google/callback'
- */
-
-// const GOOGLE_CLIENT_ID = '457946987395-f5hred2p8vbv6jsbgi60hqa3ji4l9oje.apps.googleusercontent.com'
-// const GOOGLE_CLIENT_SECRET = 'weB2LQErxwzy_OktmyA0wwfU'
-// const GOOGLE_CALLBACK = 'http://localhost:8080/auth/google/verify'
-
-if (!process.env.GOOGLE_CLIENT_ID || !process.env.GOOGLE_CLIENT_SECRET) {
-
-  console.log('Google client ID / secret not found. Skipping Google OAuth.')
-
-} else {
-
-  const googleConfig = {
-    clientID: process.env.GOOGLE_CLIENT_ID,
-    clientSecret: process.env.GOOGLE_CLIENT_SECRET,
-    callbackURL: process.env.GOOGLE_CALLBACK
-  }
-
-  const strategy = new GoogleStrategy(googleConfig, (token, refreshToken, profile, done) => {
-    const googleId = profile.id
-    const name = profile.displayName
-    const email = profile.emails[0].value
-
-    User.find({where: {googleId}})
-      .then(foundUser => (foundUser
-        ? done(null, foundUser)
-        : User.create({name, email, googleId})
-          .then(createdUser => done(null, createdUser))
-      ))
-      .catch(done)
-  })
-
-  passport.use(strategy)
-
-  router.get('/', passport.authenticate('google', {scope: ['email', 'https://www.googleapis.com/auth/fitness.activity.read', 'https://www.googleapis.com/auth/fitness.activity.write', 'https://www.googleapis.com/auth/fitness.location.read']}))
-
-  router.get('/verify', passport.authenticate('google', {
-    successRedirect: '/home',
-    failureRedirect: '/login'
-  }))
-
-}
