+++ server/auth/index.js
@@ -40,4 +40,4 @@ router.get('/me', (req, res) => {
   res.json(req.user)
 })
 
-router.use('/google', require('./google'))
+router.use('/strava', require('./strava'))
